move helpers function
