+++ trace-utils/src/test_utils/mod.rs
@@ -180,6 +180,39 @@ pub fn create_test_gcp_span(
     span
 }
 
+pub fn create_test_gcp_json_span(
+    trace_id: u64,
+    span_id: u64,
+    parent_id: u64,
+    start: i64,
+) -> serde_json::Value {
+    json!(
+            {
+                "trace_id": trace_id,
+                "span_id": span_id,
+                "service": "test-service",
+                "functionname": "dummy_function_name",
+                "name": "test_name",
+                "resource": "test-resource",
+                "parent_id": parent_id,
+                "start": start,
+                "duration": 5,
+                "error": 0,
+                "meta": {
+                    "service": "test-service",
+                    "env": "test-env",
+                    "runtime-id": "test-runtime-id-value",
+                    "gcrfx.project_id": "dummy_project_id",
+                    "_dd.mini_agent_version": "dummy_version",
+                    "_dd.gcrfx.resource_name": "projects/dummy_project_id/locations/dummy_region_west/functions/dummy_function_name",
+                    "gcrfx.location": "dummy_region_west"
+                },
+                "metrics": {},
+                "meta_struct": {},
+            }
+        )
+}
+
 pub fn create_test_json_span(
     trace_id: u64,
     span_id: u64,
